+++ sampledb/models/versioned_json_object_tables.py
@@ -249,7 +249,7 @@ class VersionedJSONSerializableObjectTables(object):
             return None
         return self.object_type(*current_object)
 
-    def get_current_objects(self, filter_func=lambda data: True, action_table=None, action_filter=None, connection=None, table=None, parameters=None):
+    def get_current_objects(self, filter_func=lambda data: True, action_table=None, action_filter=None, connection=None, table=None, parameters=None, sorting_func=None, limit=None, offset=None, num_objects_found=None):
         """
         Queries and returns all objects matching a given filter.
 
@@ -276,7 +276,8 @@ class VersionedJSONSerializableObjectTables(object):
             table.c.data,
             table.c.schema,
             table.c.user_id,
-            table.c.utc_datetime
+            table.c.utc_datetime,
+            db.sql.expression.text('COUNT(*) OVER()')
         ])
 
         if action_table is not None and action_filter is not None:
@@ -290,12 +291,30 @@ class VersionedJSONSerializableObjectTables(object):
                 )
             )
 
+        if sorting_func is None:
+            def sorting_func(columns):
+                return db.sql.desc(columns.object_id)
+
+        select_statement = select_statement.where(filter_func(table.c.data))
+        select_statement = select_statement.order_by(sorting_func(table.c))
+
+        if limit is not None:
+            select_statement = select_statement.limit(limit)
+
+        if offset is not None:
+            select_statement = select_statement.offset(offset)
+
         objects = connection.execute(
-            select_statement
-            .where(filter_func(table.c.data)),
+            select_statement,
             **parameters
         ).fetchall()
-        return [self.object_type(*obj) for obj in objects]
+        if num_objects_found is not None:
+            num_objects_found.clear()
+            if objects:
+                num_objects_found.append(objects[0][-1])
+            else:
+                num_objects_found.append(0)
+        return [self.object_type(*obj[:-1]) for obj in objects]
 
     def get_object_versions(self, object_id, connection=None):
         """